# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api eccomerce",
        "contact": {
            "email": "guilhermeieski@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuarios/{empresa_id}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "User",
                "operationId": "f8f5cd7556752de0b1503a07c1a3d1ea",
                "parameters": [
                    {
                        "name": "empresa_id",
                        "in": "path",
                        "description": "ID da empresa cujos pedidos serão buscados",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/usuario/criar": {
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Criar Usuário",
                "operationId": "9c1b0fc79fdc2048ec2e01e0ba4b6592",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "endereco_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": 11111111111
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "4792801006"
                                    },
                                    "empresa_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_master": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/usuario/{id}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "User show",
                "operationId": "3578cec20dc77ec76383f2a46f907876",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to show",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/usuario/delete/{id}": {
            "delete": {
                "tags": [
                    "Usuário"
                ],
                "summary": "User delete",
                "description": "Delete user by id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "0c3ecfce7fe2d8a8098b2f1a48272429",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
...
}